requirements.txt
streamlit==1.29.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
llama-index==0.9.27
pandas==2.1.4
numpy==1.22.4
qdrant-client==1.7.0
sentence-transformers==2.2.2
pydantic==2.5.2
typing-extensions==4.8.0
